% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-oo-r.R
\name{createObject}
\alias{createObject}
\alias{createObject}
\title{createObject}
\usage{
createObject(constructorName, ...)

createObject(constructorName, ...)
}
\arguments{
\item{constructorName}{}

\item{...}{}

\item{constructorName}{}

\item{...}{}
}
\description{
Creates functional object

createObject
}
