% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emptyList}
\alias{emptyList}
\title{Empty list}
\usage{
emptyList()
}
\description{
Empties the entire container of functional objects. Does not delete R objects
}
\seealso{
removeMember
}
