% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-oo-r.R
\name{eval.deriv.direct}
\alias{eval.deriv.direct}
\title{Evaluate derivative}
\usage{
eval.deriv.direct(addr, x, coefs, checkValidity = TRUE)
}
\arguments{
\item{addr}{address of object}

\item{x}{evaulation points}

\item{coefs}{coefficients to be multiplied. Optional}

\item{checkValidity.}{Before evaluating, check if object is valid. Changing this is completely at your own risk!}
}
\description{
A eval.direct version of eval.deriv
}
