% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-oo-r.R
\name{eval.deriv}
\alias{eval.deriv}
\title{Evaluate derivative}
\usage{
eval.deriv(fObject, x, coefs, checkValidity = TRUE)
}
\arguments{
\item{fObject}{}

\item{x}{evaulation points}

\item{coefs}{coefficients to be multiplied. Optional}

\item{checkValidity.}{Before evaluating, check if object is valid. Changing this is completely at your own risk!}
}
\description{
Evaluate derivative
}
