% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-oo-r.R
\name{eval.direct}
\alias{eval.direct}
\title{Evaluates functional object}
\usage{
eval.direct(address, x, coefs, checkValidity = TRUE)
}
\arguments{
\item{address}{address}

\item{x}{points in which to evaluate}

\item{coefs}{coefficients to be multiplied. Optional}

\item{checkValidity.}{Before evaluating, check if object is valid. Changing this is completely at your own risk!}
}
\value{
A matrix containing the functional basis evaluated in x,
if coefs are missing. Otherwise a vector containg evaluation in x with said coefficients.
}
\description{
As eval.fct
}
