% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctbases.R
\docType{package}
\name{fctbases-package}
\alias{fctbases}
\alias{fctbases-package}
\title{fctbases: Functional bases}
\description{
Easy-to-use, very fast implementations of various functional bases
}
\details{
fctbases is a fast and easy implementation of functional bases in R. Simply initialize the desired basis, which returns function of class \code{fctbases}.

Internally, functions are stored as C++ objects are stored in C++.
}
