% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_fourier_basis}
\alias{init_fourier_basis}
\title{Initialize Fourier basis}
\usage{
init_fourier_basis(range, f_order, j = FALSE)
}
\arguments{
\item{range}{Left and right end points.}

\item{spline_order}{order = degree + 1}

\item{nknots}{No. of knots including end points}
}
\description{
Assign value to the global b-spline.
}
