% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_unif_bspline}
\alias{init_unif_bspline}
\title{Initialize uniform B-spline}
\usage{
init_unif_bspline(range, nknots, spline_order, ind_count = FALSE)
}
\arguments{
\item{range}{Left and right end points.}

\item{nknots}{No. of knots including end points}

\item{spline_order}{order = degree + 1}
}
\description{
Initilaises a uniform b-spline.
}
