% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initBspline}
\alias{initBspline}
\title{Initialize B-spline}
\usage{
initBspline(spline_order, spline_knots, ind_count = FALSE)
}
\arguments{
\item{spline_order}{order = degree + 1}

\item{spline_knots}{Knots of the b-spline, including endpoints. Must be increasing.}
}
\value{
integer.
}
\description{
Initialises b-spline and returns address.
}
