% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initBspline1}
\alias{initBspline1}
\title{Initialize B-spline of order one}
\usage{
initBspline1(spline_knots, ind_count = FALSE)
}
\arguments{
\item{spline_knots}{Knots of the b-spline, including endpoints. Must be increasing.}
}
\value{
integer.
}
\description{
Initialize B-spline of order one
}
