% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make.bspline.basis}
\alias{make.bspline.basis}
\title{Make b-spline basis}
\usage{
make.bspline.basis(knots, order = 4)
}
\arguments{
\item{knots}{Knots of the basis, including endpoints}

\item{order}{Spline order. Defaults to 4.}
}
\value{
Function
}
\description{
Make b-spline basis
}
\seealso{
\link{Functional basis function}
}
