% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-oo-r.R
\name{make.bspline}
\alias{make.bspline}
\title{Make b-spline}
\usage{
make.bspline(type = c("standard"), knots = NULL, range = c(0, 1),
  order = 4, nknots = NULL)
}
\arguments{
\item{type}{"standard" or "uniform" available}

\item{knots}{For type = "standard", specify knots. Must be increasing.}

\item{range}{For type = "uniform", endpoints of spline.}

\item{order}{Order of bspline. If type == standard and order == 1, bspline1 is used.}

\item{nknots}{For type = "uniform", no. of knots (endpoints included).}
}
\value{
A functionObject with some info.
}
\description{
Make b-spline
}
