% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make.fourier.basis}
\alias{make.fourier.basis}
\title{Make fourier basis}
\usage{
make.fourier.basis(range, order, use.trig.id = FALSE)
}
\arguments{
\item{range}{Left and right end points.}

\item{order}{Order of harmonics}

\item{use.trig.id}{Use trigonometrical identities with this function?}
}
\value{
Function
}
\description{
Make fourier basis
}
\details{
The number of basis elements (degrees of freedom) is 2 * order + 1.

The basis functions are  [1, sin(x), cos(x), sin(2x), cos(2x), ..., sin(nx), cos(nx) ]
}
