% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make.std.bspline.basis}
\alias{make.std.bspline.basis}
\title{'Standard' B-spline basis}
\usage{
make.std.bspline.basis(range, intervals)
}
\arguments{
\item{range}{End points of spline}

\item{intervals}{Number of intervals}
}
\value{
function
}
\description{
This initializes a bspline of order 4 with uniformly places knots. df = intervals + 3.
}
