% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctbases.R
\name{object.info}
\alias{object.info}
\title{Functional basis info}
\usage{
object.info(fctbasis)
}
\arguments{
\item{fctbasis}{object of class \code{fctbasis}}
}
\value{
A named list including no. of basis, type of basis, and possibly additional information.
}
\description{
This function returns details about a functional basis.
}
