% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeMember}
\alias{removeMember}
\title{Removes a functional object from the list.}
\usage{
removeMember(address)
}
\description{
Returns true if member awas on the list, returns false otherwise.
}
