% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctbases.R
\docType{package}
\name{fctbases-package}
\alias{fctbases}
\alias{fctbases-package}
\title{fctbases: Functional bases}
\description{
fctbases is a fast and easy implementation of functional bases in R. Simply initialize the desired basis, which returns function of class \code{fctbasis}.
}
\details{
Internally, functions are stored as C++ objects, which are masked by the package.
The package maintains the bookkeeping of fctbasis objects. Parameters are validated at initialization which also reduces some of the overhead.
fctbases objects cannot be saved across sessions and must be re-initialised.

Derivatives are provided. These are the mathematically correct ones and are as fast as the non-derivatives.
}
\seealso{
\link{Functional basis function}
}
