% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctbases.R
\name{Functional basis function}
\alias{Functional basis function}
\title{Functional basis function}
\arguments{
\item{t}{time points}

\item{x}{vector of coefficients (optional)}

\item{deriv}{Should the derivative be used and which order? Defaults to \code{FALSE}}
}
\value{
Returns a matrix if x is missing, and a vector if x is provided.
}
\description{
A fctbases object is a function of class \code{fctbases} which takes three arguments \code{(t, x, deriv)}
}
\details{
If \code{deriv} is \code{FALSE} or zero, the function itself is evaluated.
If \code{deriv} is one or \code{TRUE}, the first derivative is evaluated.
If \code{deriv} is two, the second derivative is evaluated.
}
