% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_fourier_basis}
\alias{init_fourier_basis}
\title{Init fouriter basis}
\usage{
init_fourier_basis(range, order, trig_basis = FALSE)
}
\arguments{
\item{trig_basis}{Currently ignored.}
}
\description{
Init fouriter basis
}
