% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make.fourier.basis}
\alias{make.fourier.basis}
\title{Make fourier basis}
\usage{
make.fourier.basis(range, order, df, mode = c("function", "pointer",
  "f_object"), use.trig.id = FALSE)
}
\arguments{
\item{range}{Left and right end points.}

\item{df}{Not used}

\item{mode}{Not used}
}
\value{
Function
}
\description{
Make fourier basis
}
\details{
The number of base elements is 2 * order + 1.
}
