% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make.pol.basis}
\alias{make.pol.basis}
\title{Make polynomial basis}
\usage{
make.pol.basis(order, mode = c("function", "pointer", "f_object"),
  ty = F)
}
\arguments{
\item{order}{Order of polynomial (= degree + 1)}
}
\value{
Function
}
\description{
Make polynomial basis
}
\details{
Polynomial basis (1, x, x^2, x^3, ..., x^n)
}
