% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make.std.bspline.basis}
\alias{make.std.bspline.basis}
\title{'Standard' B-spline basis}
\usage{
make.std.bspline.basis(range = c(0, 1), intervals)
}
\arguments{
\item{range}{End points of spline}

\item{intervals}{Number of intervals}
}
\value{
function
}
\description{
This initializes a bspline of order 4 with uniformly places knots. df = intervals + 3.
}
\details{
\code{make.std.bspline.basis} uses a different implementation than \code{make.bspline.basis}, 
but is not faster in all uses.
}
\examples{
## 16 equidistant knots between 0 and 2 (both included)
bf <- make.std.bspline.basis(range = c(0,2), intervals = 15)

}
\seealso{
\link{Functional basis function}, \link{make.bspline.basis}
}
