% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeMember}
\alias{removeMember}
\title{Deletes a functional object from the list of available objects.}
\usage{
removeMember(address)
}
\description{
Returns true if the member correctly was on the list, returns false otherwise.
}
